Address subprocess completion review feedback
