+++ AGENTS.md
@@ -118,7 +118,7 @@ This module is **pre-spawn planning only**: it validates raw `args`, derives imp
 - **Launch-scoped flags** — `getStartupScopedFlags` mirrors `LAUNCH_SCOPED_FLAG_DEFINITIONS` (`--auto-connect`, `--cdp`, `--enable`, `--init-script`, `--profile`, `--session-name`, `--state`, `--provider`, `-p`, `--device`). `--auto-connect` counts only when enabled (bare or `true`); `--state` after a `wait` command is treated as a wait predicate, not launch state, so it does not force `sessionMode: "fresh"` during implicit reuse.
 - **Tab-correction signal** — `hasLaunchScopedTabCorrectionFlag` is a subset (`--profile`, `--session-name`, `--state`) for post-open tab re-selection when restored tabs steal focus; it intentionally excludes `--cdp`, `--auto-connect`, and provider selection flags (`-p`, `--provider`, `--device`).
 - **Upstream subprocess environment** — `extensions/agent-browser/lib/process.ts` passes through a curated env: always-inherited safe names plus every `AGENT_BROWSER_*` variable and provider-related prefixes (`AGENTCORE_`, `BROWSERBASE_`, `BROWSERLESS_`, `BROWSER_USE_`, `KERNEL_`) alongside `AI_GATEWAY_*` and `XDG_*`; extend that module when new upstream credential env families must reach the child without leaking unrelated parent secrets.
-- **Upstream subprocess completion** — `runAgentBrowserProcess` spawns `agent-browser` with `stdio: ["pipe", "pipe", "pipe"]` and must not wait forever on `"close"` when a detached descendant inherits those streams. `watchSpawnedChildCompletion` listens for `exit` and `close`, destroys piped stdio on `exit` to drop inherited handles promptly, then applies a short post-`exit` grace (`EXIT_STDIO_GRACE_MS`) before resolving with the observed exit code when `close` is still delayed. Exit-code precedence is `close` → wrapper timeout (`124`) → post-`exit` fallback → spawn failure (`127`), locked by `resolveSpawnedChildExitCode` and `test/agent-browser.process.test.ts`. Changing that window, precedence, or stream teardown order requires updating [`docs/ARCHITECTURE.md`](docs/ARCHITECTURE.md) subprocess bullets, [`docs/SUPPORT_MATRIX.md`](docs/SUPPORT_MATRIX.md) `RQ-0097`, and the stdio-linger regressions in `test/agent-browser.process.test.ts`.
+- **Upstream subprocess completion** — `runAgentBrowserProcess` spawns `agent-browser` with `stdio: ["pipe", "pipe", "pipe"]` and must not wait forever on `"close"` when a detached descendant inherits those streams. `watchSpawnedChildCompletion` listens for `exit` and `close`, leaves piped stdio intact for the short post-`exit` grace (`EXIT_STDIO_GRACE_MS`) so normal `close` can still win, then destroys those streams only if the fallback resolves with the observed `exit` code. Exit-code precedence is `close` → wrapper timeout (`124`) → post-`exit` fallback → spawn failure (`127`), locked by `resolveSpawnedChildExitCode` and `test/agent-browser.process.test.ts`. Changing that window, precedence, or stream teardown order requires updating [`docs/ARCHITECTURE.md`](docs/ARCHITECTURE.md) subprocess bullets, [`docs/SUPPORT_MATRIX.md`](docs/SUPPORT_MATRIX.md) `RQ-0097`, and the stdio-linger regressions in `test/agent-browser.process.test.ts`.
 - **Stateless inspection** — `isPlainTextInspectionArgs` recognizes `--help`, `-h`, `--version`, and `-V`: those calls keep the caller’s argv (no `--json`), skip implicit `--session`, and are ignored when `restoreManagedSessionStateFromBranch` walks prior tool rows. Read-only `skills list`, `skills get …`, and `skills path …` are treated as stateless JSON inspection via `isStatelessInspectionCommand`: `buildExecutionPlan` still prepends `--json` but skips the same implicit-session and fresh-session injection paths as plain-text inspection, so successful results omit managed `sessionName` / `usedImplicitSession` and do not advance session replay. Any new upstream `skills` subcommand only picks up this behavior after it is allowlisted next to `isStatelessInspectionCommand` and covered in `test/agent-browser.runtime.test.ts` (see `buildExecutionPlan limits stateless skills inspection to documented subcommands`).
 - **Regression tests** — `test/agent-browser.runtime.test.ts` locks execution-plan injection, launch-scoped blocking, argv edge cases, and related helpers.
